requirements.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
click-plugins==1.1.1
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contextlib2==21.6.0
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cupy-cuda12x==13.3.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
docutils==0.20.1
fastrlock==0.8.2
fonttools==4.46.0
fs==2.4.16
gps==3.25
html5lib==1.1
httplib2==0.20.4
idna==3.6
imagesize==1.4.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
mpmath==0.0.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
packaging==24.0
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
py==1.11.0
pycairo==1.25.1
pycups==2.0.1
pygccxml==2.4.0
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyOpenGL==3.1.7
pyparsing==3.1.1
PyQt-Qwt==1.2.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQt6==6.6.1
PyQt6-sip==13.6.0
pyqtgraph==0.13.4
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.31.0
rich==13.7.1
roman==3.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
schema==0.7.5
SciPy==1.11.4
screen-resolution-extra==0.0.0
setuptools==68.1.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
ssh-import-id==5.11
sshuttle==1.1.1
sympy==1.12
systemd-python==235
thrift==0.19.0
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
wadllib==1.3.6
webencodings==0.5.1
wheel==0.42.0
xdg==5
xkit==0.0.0
